feat(Server): on the server add the Straight pattern the ennemy with the straight pattern move in line close #95
