+++ src/shared/components/StraightPattern.hpp
@@ -1,0 +1,19 @@
+/*
+** EPITECH PROJECT, 2024
+** R-Type-epitech
+** File description:
+** StraightPattern
+*/
+
+#ifndef STRAIGHT_PATTERN_HPP_
+    #define STRAIGHT_PATTERN_HPP_
+
+    typedef struct StraightLinePattern_s {
+        /**
+         * @brief StraightLinePattern add a StraightLine patern for an entity
+         *
+         */
+        float speed;
+    } StraightLinePattern;
+
+#endif
