+++ src/shared/components/Wave_Pattern.hpp
